исправлено кой-чего
ы

+++ Palitra.h
@@ -23,11 +23,11 @@ public:
 		green = min + (max - min) / 2;
 		yellow = max - (max - min) / 4;;
 		blue = min + (max - min) / 4;;
-		Red = to_string((int)red);
-		Blue = to_string((int)blue);
-		Yellow = to_string((int)yellow);
-		Green = to_string((int)green);
-		Minblue = to_string((int)minblue);
+		Red = to_string(red);
+		Blue = to_string(blue);
+		Yellow = to_string(yellow);
+		Green = to_string(green);
+		Minblue = to_string(minblue);
 	}
 
 	void color(int &r, int &g, int &b, double value)
